requirements.txt
streamlit
torch
torchvision
numpy
matplotlib
pillow
gdown
